bugfix: unify for self containing less nodes than other
Function to insert missing rows not properly handling the case where self's
nodes are a subset of other's nodes. Previous implementation was only adding a
new _missing_nodes column if self's nodes were outside the set of other's
nodes, causing NaNs to be defined in the column and incorrect tree printout
behavior.